Move logger and add warning if the chennel is full
